A bit more cleanup, removed some redundant files. Last night I think I coupled the preparsing functions too heavily with the PreparseData struct (i.e. made them apply parsed data directly to struct rather than return them), and am going to probably change that back.
