+++ compiler/cx-compiler-ast/src/lib.rs
@@ -3,7 +3,6 @@ mod log;
 
 pub(crate) mod declarations;
 pub(crate) mod definitions;
-pub(crate) mod parsing_tools;
 
 pub mod parse;
 pub mod preparse;
